tools: rename BaseTool.py->base_tool.py, ToolFactory.py->tool_factory.py; update imports and tests; add plan progress
